Basic implementation done

+++ Qt-task-1/polarFuncLib/archspiralfunc.cpp
@@ -7,7 +7,7 @@ ArchSpiralFunc::ArchSpiralFunc()
 
 }
 
-float ArchSpiralFunc::eval(float theta)
+double ArchSpiralFunc::eval(double theta)
 {
     return 0.1*theta;
 }
