+++ Qt-task-1/polarFuncLib/archspiralfunc.h
@@ -12,7 +12,7 @@ private:
 public:
     ArchSpiralFunc();
 
-    float eval(float theta) override;
+    double eval(double theta) override;
     std::string getName() const override;
 };
